requirements.txt
langchain-core==0.2.33
langgraph==0.2.4
langchain-experimental==0.0.64
langchain-community==0.2.12
langchain-openai==0.1.22
langchain-anthropic==0.1.23
beautifulsoup4==4.12.3
termcolor==2.4.0
chainlit==1.1.202
colorlog==6.8.2
fake-useragent==1.5.1
playwright==1.45.0
pypdf==4.2.0
llmsherpa==0.1.4
fastembed==0.3.4
faiss-cpu==1.8.0.post1
FlashRank==0.2.6
chromadb==0.5.5
timeout-decorator==0.5.0
neo4j==5.23.1
langsmith==0.1.129

# Microservice dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.7.4
aiohttp>=3.9.1
httpx>=0.25.2
python-multipart>=0.0.6
PyYAML>=6.0.1

# syncer==2.0.3
